+++ lib/Colfeed.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:nie/Card.dart';
 import 'package:nie/globalvariables.dart';
 
 class Colfeed extends StatefulWidget {
@@ -37,33 +38,7 @@ class _ColfeedState extends State<Colfeed> {
 					],
 				)
 			),
-			Padding(
-				padding: EdgeInsets.all(12),
-				child: Card(
-					child: Column(
-						mainAxisSize: MainAxisSize.min,
-						children: <Widget>[
-							const ListTile(
-								leading: Icon(Icons.album),
-								title: Text('The Enchanted Nightingale'),
-								subtitle: Text('Music by Julie Gable. Lyrics by Sidney Stein.'),
-							),
-							ButtonBar(
-								children: <Widget>[
-									FlatButton(
-										child: const Text('BUY TICKETS'),
-										onPressed: () { /* ... */ },
-									),
-									FlatButton(
-										child: const Text('LISTEN'),
-										onPressed: () { /* ... */ },
-									),
-								],
-							),
-						],
-					),
-				),
-			)
+			FeedCard(),
 		],
 	);
   }
